# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ScriptForge API
  version: 1.0.0
  description: ScriptForge 前台 / 后台 REST API 契约
paths:
...
  /api/creation/projects/{project_id}/agents/{node_index}/content/:
    put:
      operationId: creation_projects_agents_content_update
      description: PUT /api/creation/projects/<project_id>/agents/<node_index>/content/
      parameters:
      - in: path
        name: node_index
        schema:
          type: integer
        required: true
      - in: path
        name: project_id
        schema:
          type: string
        required: true
      tags:
      - creation
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          description: No response body
  /api/creation/projects/{project_id}/agents/{node_index}/generate/:
    post:
      operationId: creation_projects_agents_generate_create
      description: POST /api/creation/projects/<project_id>/agents/<node_index>/generate/
      parameters:
      - in: path
        name: node_index
        schema:
          type: integer
        required: true
      - in: path
        name: project_id
        schema:
          type: string
        required: true
      tags:
      - creation
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          description: No response body
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT